# api/swagger.yaml
...
swagger: "2.0"
info:
  version: "1.0.0"
  title: "Jobjoo API Documentation"
host: "localhost:4000"
basePath: "/"
...
schemes:
- "http"
paths:
...
  /resume-creator/step-five:
    get:
      tags:
      - "resume-creator"
      summary: "Get Step Five Data"
      operationId: "getResumecreatorStepfive"
      parameters: []
      responses:
        default:
          description: "Successful"
          schema:
            type: "string"
      x-swagger-router-controller: "ResumeCreator"
    post:
      tags:
      - "resume-creator"
      summary: "Post Step Five Data"
      operationId: "postResumecreatorStepfive"
      parameters:
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/Model 19"
      responses:
        default:
          description: "Successful"
          schema:
            type: "string"
      x-swagger-router-controller: "ResumeCreator"
  /resume-creator/step-four:
    get:
      tags:
      - "resume-creator"
      summary: "Get Step Four Data"
      operationId: "getResumecreatorStepfour"
      parameters: []
      responses:
        default:
          description: "Successful"
          schema:
            type: "string"
      x-swagger-router-controller: "ResumeCreator"
    post:
      tags:
      - "resume-creator"
      summary: "Post Step Four Data"
      operationId: "postResumecreatorStepfour"
      parameters:
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/Model 22"
      responses:
        default:
          description: "Successful"
          schema:
            type: "string"
      x-swagger-router-controller: "ResumeCreator"
  /resume-creator/step-one:
    get:
      tags:
      - "resume-creator"
      summary: "Get Step One Data"
      operationId: "getResumecreatorStepone"
      parameters: []
      responses:
        default:
          description: "Successful"
          schema:
            type: "string"
      x-swagger-router-controller: "ResumeCreator"
    post:
      tags:
      - "resume-creator"
      summary: "Post Step One Data"
      operationId: "postResumecreatorStepone"
      parameters:
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/Model 23"
      responses:
        default:
          description: "Successful"
          schema:
            type: "string"
      x-swagger-router-controller: "ResumeCreator"
  /resume-creator/step-three:
    get:
      tags:
      - "resume-creator"
      summary: "Get Step Three Data"
      operationId: "getResumecreatorStepthree"
      parameters: []
      responses:
        default:
          description: "Successful"
          schema:
            type: "string"
      x-swagger-router-controller: "ResumeCreator"
    post:
      tags:
      - "resume-creator"
      summary: "Post Step Three Data"
      operationId: "postResumecreatorStepthree"
      parameters:
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/Model 25"
      responses:
        default:
          description: "Successful"
          schema:
            type: "string"
      x-swagger-router-controller: "ResumeCreator"
  /resume-creator/step-two:
    get:
      tags:
      - "resume-creator"
      summary: "Get Step Two Data"
      operationId: "getResumecreatorSteptwo"
      parameters: []
      responses:
        default:
          description: "Successful"
          schema:
            type: "string"
      x-swagger-router-controller: "ResumeCreator"
    post:
      tags:
      - "resume-creator"
      summary: "Post Step Two Data"
      operationId: "postResumecreatorSteptwo"
      parameters:
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/Model 27"
      responses:
        default:
          description: "Successful"
          schema:
            type: "string"
      x-swagger-router-controller: "ResumeCreator"
...
  /search/complex/jobs:
    get:
      tags:
      - "search"
      summary: "Return A List Of Jobs Based On Query [Complex Search]"
      operationId: "getSearchComplexJobs"
      parameters:
      - name: "q"
        in: "query"
        required: true
        type: "string"
        maxLength: 30
      - name: "contractTypes"
        in: "query"
        required: false
        type: "string"
      responses:
        default:
          description: "Successful"
          schema:
            type: "string"
      x-swagger-router-controller: "Search"
  /search/simple/jobs:
    get:
      tags:
      - "search"
      summary: "Return A List Of Jobs Based On Query [Simple Search]"
      operationId: "getSearchSimpleJobs"
      parameters:
      - name: "q"
        in: "query"
        required: true
        type: "string"
        maxLength: 30
      responses:
        default:
          description: "Successful"
          schema:
            type: "string"
      x-swagger-router-controller: "Search"
...
  /register/register-number:
    post:
      tags:
      - "register"
      summary: "Register Phone Number"
      operationId: "postRegisterRegisternumber"
      parameters:
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/Model 15"
      responses:
        default:
          description: "Successful"
          schema:
            type: "string"
      x-swagger-router-controller: "Register"
  /register/validate-number:
    post:
      tags:
      - "register"
      summary: "Validate Registered Number"
      operationId: "postRegisterValidatenumber"
      parameters:
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/Model 16"
      responses:
        default:
          description: "Successful"
          schema:
            type: "string"
      x-swagger-router-controller: "Register"
  /register/validate-number-set-user:
    post:
      tags:
      - "register"
      summary: "Validate Registered Number and set user "
      operationId: "postRegisterValidatenumbersetuser"
      parameters:
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/Model 17"
      responses:
        default:
          description: "Successful"
          schema:
            type: "string"
      x-swagger-router-controller: "Register"
...
securityDefinitions:
  jwt:
    type: "apiKey"
    name: "Authorization"
    in: "header"
...